+++ packages/literature_core/lib/picklogic_literature_core.dart
@@ -1,10 +1,15 @@
 library;
 
 export 'src/citation_export.dart';
+export 'src/citation_import.dart';
+export 'src/citation_styles.dart';
 export 'src/doi_extractor.dart';
 export 'src/library_catalog.dart';
+export 'src/library_organization.dart';
 export 'src/literature_annotation.dart';
 export 'src/literature_naming.dart';
 export 'src/pdf_metadata_reader.dart';
 export 'src/reading_progress.dart';
+export 'src/reference_duplicates.dart';
+export 'src/translation_memory.dart';
 export 'src/translation_provider.dart';